Struct clarity_lsp::clarity::analysis::read_only_checker::ReadOnlyChecker

pub struct ReadOnlyChecker<'a, 'b> {
    db: &'a mut AnalysisDatabase<'b>,
    defined_functions: HashMap<ClarityName, bool>,
}

Fields
db: &'a mut AnalysisDatabase<'b>defined_functions: HashMap<ClarityName, bool>
Methods

impl<'a, 'b> ReadOnlyChecker<'a, 'b>

fn new(db: &'a mut AnalysisDatabase<'b>) -> ReadOnlyChecker<'a, 'b>

pub fn run(
&mut self,
contract_analysis: &mut ContractAnalysis
) -> CheckResult<()>

fn check_define_function(
&mut self,
signature: &[SymbolicExpression],
body: &SymbolicExpression
) -> CheckResult<(ClarityName, bool)>

fn check_reads_only_valid(
&mut self,
expr: &SymbolicExpression
) -> CheckResult<()>

fn check_read_only(&mut self, expr: &SymbolicExpression) -> CheckResult<bool>

Checks the supplied symbolic expressions (1) for whether or not they are valid with respect to read-only requirements. (2) if valid, returns whether or not they are read only. Note that because of (1), this function cannot short-circuit on read-only.

fn check_all_read_only(
&mut self,
expressions: &[SymbolicExpression]
) -> CheckResult<bool>

Checks all of the supplied symbolic expressions (1) for whether or not they are valid with respect to read-only requirements. (2) if valid, returns whether or not they are read only. Note that because of (1), this function cannot short-circuit on read-only.

fn is_implicit_tuple_definition_read_only(
&mut self,
tuples: &[SymbolicExpression]
) -> CheckResult<bool>

fn try_native_function_check(
&mut self,
function: &str,
args: &[SymbolicExpression]
) -> Option<CheckResult<bool>>

fn check_native_function(
&mut self,
function: &NativeFunctions,
args: &[SymbolicExpression]
) -> CheckResult<bool>

fn check_function_application_read_only(
&mut self,
expression: &[SymbolicExpression]
) -> CheckResult<bool>